Fix sticker resizing.
So far, a sticker of size 1028x1028 would have been resized to 128x128,
even though the maximum was set to 256x256.
This fix introduces a different way of resizing, which means that any dimension
of length >256 will result =256.

Also, I factored out the maximum size, so maybe we can make this an option in the future.