Override dh_auto_clean
It attempts to call 'make distclean', which appears to be a sane default.
However, telegram-purple does not have a concept of distclean, and we're
unwilling to introduce one only for Debian. So just go for the target 'clean'
instead, which currently does to right thing, which is: remove *everything*
that can be regenerated during build.